requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymupdf==1.24.3
langchain-text-splitters==0.2.0
openai==1.30.0
qdrant-client==1.9.1
python-multipart==0.0.9
python-dotenv==1.0.1
